go-backup/requirements.txt
dify-plugin-sdk>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python3-saml>=1.16.0
authlib>=1.3.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
structlog>=24.1.0
cryptography>=42.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
bandit>=1.7.0
safety>=3.0.0 